+++ sdk/resourcemanager/hybridnetwork/armhybridnetwork/ze_generated_example_vendorskus_client_test.go
@@ -184,7 +184,7 @@ func ExampleVendorSKUsClient_BeginCreateOrUpdate() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/hybridnetwork/resource-manager/Microsoft.HybridNetwork/preview/2022-01-01-preview/examples/VendorSkuListByVendor.json
-func ExampleVendorSKUsClient_List() {
+func ExampleVendorSKUsClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -196,7 +196,7 @@ func ExampleVendorSKUsClient_List() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.List("<vendor-name>",
+	pager := client.NewListPager("<vendor-name>",
 		nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)